chore: move files out of util folder
