+++ src/encoded/static/components/viz/AboutPageVisualizations/Alluvial/data/alluvial_data.json
@@ -374,45 +374,12 @@
             "Tranquil-Seq": [ "PacBio" ]
         }
     },
-    "map": {
-        "BCM": 0,
-        "Broad": 1,
-        "NYGC": 2,
-        "UW": 3,
-        "WashU": 4,
-        "Illumina NovaSeq": 5,
-        "PacBio": 6,
-        "Ultima Genomics": 7,
-        "ONT": 8,
-        "ONT-Ultra-Long": 9,
-        "10X Genomics Xenium": 10,
-        "WGS": 11,
-        "WGS DLP+": 12,
-        "WGS Duplex sequencing": 13,
-        "WGS CODEC": 14,
-        "WGS Nanoseq": 15,
-        "NT-Seq": 16,
-        "Hi-C": 17,
-        "RNA-Seq": 18,
-        "STORM-Seq": 19,
-        "Fiber-Seq": 20,
-        "snRNA-Seq": 21,
-        "Tranquil-Seq": 22,
-        "MAS-Seq": 23,
-        "Targeted RNA-Seq": 24,
-        "SNV": 25,
-        "InDel": 26,
-        "CNV": 27,
-        "SV": 28,
-        "MEI": 29,
-        "Repeat elements": 30,
-        "Telomere lengths": 31,
-        "Assembled genome": 32,
-        "DNA methylation": 33,
-        "Chromatin accessibility": 34,
-        "Chromatin conformation": 35,
-        "Gene expression": 36,
-        "Transcript isoform": 37,
-        "Gene expression - Spatial": 38
+    "colors": {
+        "data_generator": ["#39A2AC", "#BCBD22", "#7F7F7F", "#DE7154", "#8C564B"],
+        "sequencing_platform": ["#9467BD", "#D62728", "#AFE35F", "#E78026", "#1F77B4", "#E377C2"],
+        "assay_type": ["#96897B", "#8CBBD9", "#A87F38", "#4DCE46", "#E2E050",  "#DEB487", "#C464C3", "#3D348B", "#FFD766", "#134074", "#698996", "#DAD6D6", "#CEB3AB", "#CF995F", "#E87D91"],
+        "genetic": ["#AFA8F5"],
+        "epigenetic": ["#AD4A66"],
+        "transcriptomic": ["#3D8F3D"]
     }
 }
